pyproject.toml
[tool.poetry]
name = "pygame_qt"
version = "0.1.0"
description = ""
authors = ["Violet Eldridge <github@violet-mica.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pyside6 = "^6.7.2"
pygame-ce = {path = "../../../pygame-ce"}


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
pytest = "^8.3.2"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
